### 2019-09-07 #341 @sckott
submission: parzer
Submitting Author: Scott Chamberlain (@sckott)
Repository:  https://github.com/ropenscilabs/parzer
Version submitted:  `v0.0.4.9110`
Version accepted: TBD

---

```
Package: parzer
Title: Parse Coordinates
Description: Parse Coordinates.
Version: 0.0.4.9110
		person("Scott", "Chamberlain", role = c("aut", "cre"),
            email = "sckott@protonmail.com")
	)
    Rcpp (>= 1.0.2)
    roxygen2 (>= 6.1.1),
    randgeo,
LinkingTo: Rcpp
SystemRequirements: C++11

```

`parzer` fits most closely with the description "manipulating geospatial data" within the geospatial category - b/c it parses many variation of lat/lat coordinates.

People working with spatial data, either in academia, government, tech, etc.

Not that I know of

N/A
