### 2019-09-22 #341 @anadiedrichs
### Editor checks:

---

#### Editor comments

Hello @sckott , thanks for submitting to rOpenSci. I'll be your handling editor.

The following is the output of goodpractice::gp().

```
It is good practice to

    R/dms-fxns.R:6:NA
    R/parse_lat.R:62:NA
    R/parse_lon.R:67:NA
    R/zzz.R:20:NA

  ✖ fix this R CMD check ERROR: Package suggested but not available:
    ‘randgeo’ The suggested packages are required for a complete check. Checking can be attempted without them by setting the environment variable
    _R_CHECK_FORCE_SUGGESTS_ to a false value. See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’ manual.

```
Output of covr::package_coverage()

```
covr::package_coverage()
parzer Coverage: 93.33%
R/parse_lat.R: 75.00%
R/parse_lon.R: 75.00%
R/zzz.R: 88.89%
R/dms-fxns.R: 96.55%
R/parse_hemisphere.R: 100.00%
R/parse_lat_lon.R: 100.00%
R/parse_parts.R: 100.00%
```

Please, fix your DESCRIPTION file. Take a look at all of this, and let me know once it is done so that I can check again.

It's my first time as an invited editor. @melvidoni: should the package tests code coverage be 100%?  93 % of coverage is an excellent start point. If it isn't, can I start looking for reviewers if everything is fixed?

---
