### 2019-09-23 #341 @melvidoni
Hey @anadiedrichs the % of test is relative to the package, but 93% is great in my personal opinion. Yes, you can start looking for reviewers!
