### 2019-10-02 #341 @anadiedrichs
Hi @sckott, you are right. Consider checking your DESCRIPTION file in case you plan later to submit to CRAN to avoid problems with this issue.
I'm still seeking reviewers.
