### 2019-10-04 #341 @anadiedrichs
We have one confirmed reviewer, Maria Victoria Munafó @mvickm. She's a contributor to the water R package (https://CRAN.R-project.org/package=water) and an active R developer.

I am still looking for the second one.
