### 2019-10-15 #341 @anadiedrichs
We have another reviewer, Julien Brun @brunj7 (http://brunj7.github.io/about/)!! :-)

Thanks for your collaboration @mvickm  and @brunj7 to review this issue for rOpenSci.
You have a period of three weeks to review this package from  18th October to 8th November included.

Let me know any questions you have.
