### 2019-10-23 #341 @anadiedrichs
Hi Nicholas @njtierney. Of course you can review this issue, but first please make sure you do not have a [conflict of interest](https://ropensci.github.io/dev_guide/policies.html#coi) preventing you from reviewing this package.
