### 2019-11-08 #341 @brunj7
Hi @anadiedrichs and @sckott thank you for the opportunity to review `parzer`!  I have learned a lot from reading @sckott 's code. It is my first review for ropensci, so feel free to ask any questions about my review.

I think `parzer` is already a solid package and I can see many uses of it to help cleaning some messy data.

My 2 main comments are:

- I think it will be better if the functions using both lon and lat had their parameters switched so it matches the GIS "convention" of a point defined a c(x=lon, y=lat)
- I think adding a use case with `parzer` working with a spatial package could be very valuable to the package documentation

Here is my review:

- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be auto-generated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Suggestions / comments:

#### Description of the package

The description of the package could be more specific regarding how it refers to `coordinates`.

- Maybe `geographic coordinates` or `latitude and longitude coordinates`? I personally prefer the former because the use of geographic could also be referring to geographic vs projected coordinate systems (not covered by `parzer` by scope definition).

- Adding more information on how parzer differentiate itself and/or complement `sp::char2dms()` or `biogeo::dms2dd()` for example

-  Maybe add/switch an example to have using a Southern Hemisphere Lat using `S`? Just to demonstrate negative numbers transformation.
- Maybe adding a vignette of a use case of `parzer` of working with one of the spatial packages? I personally think that will reinforce the statement of need.

### Specific functions

- for the functions using both Latitude and Longitude. By convention, in most of the GIS software/libraries (including `sf`) coordinates are defined with `c(x=lon, y=lat)` format. I suggest adopting this format for all the functions of `parzer` using both latitude and longitude (`parse_hemisphere`, `parse_lat_long`). For those same functions, I also suggest in the error handling of out of bound values to ask the user to check if they have not inverted Lon-lat.
- I am not 100% convinced about the one letter function naming (`d()`, `m()`, `s()`)... although I have to admit I am short in suggestions as the name `degree` is probably used in many packages and won't be necessarily less ambiguous.

### Other

- Not quite sure why `goodpractice` returned `covr` test as failed, whereas the unit testing coverage of the package seems to be good. Otherwise, I ran several checks everything returned fine (https://github.com/brunj7/parzer-review/blob/master/jb_checks.R). I have seen some comments about that earlier in this thread.but

```r
covr::package_coverage()
# parzer Coverage: 93.65%
# R/parse_lat.R: 75.00%
# R/parse_lon.R: 75.00%
# R/zzz.R: 88.89%
# src/latlong.cpp: 90.69%
# R/dms-fxns.R: 96.55%
# R/parse_hemisphere.R: 100.00%
# R/parse_lat_lon.R: 100.00%
# R/parse_parts.R: 100.00%
# src/parse.cpp: 100.00%
# src/pz_hemisphere.cpp: 100.00%
# src/pz_parse_parts.cpp: 100.00%
```

- Is there a way to reduce Rccp dependencies?
