### 2019-11-09 #341 @mvickm
Hello @anadiedrichs and @sckott , I want to say thank you to consider me for the review as @brunj7  said. Also, it's my first time doing it and I've learn a lot . Thank you again. Here is my review:

## Package Review

#### Documentation

- [ x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
- [ x] **Installation instructions:** for the development version of package and any non-standard dependencies in README
- [ x] **Vignette(s)** demonstrating major functionality that runs successfully locally
- [ x] **Function Documentation:** for all exported functions in R help
- [ x] **Examples** for all exported functions in R Help that run successfully locally
- [ x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

#### Functionality

- [x ] **Installation:** Installation succeeds as documented.
- [ x] **Functionality:** Any functional claims of the software been confirmed.
- [ x] **Performance:** Any performance claims of the software been confirmed.
- [x ] **Automated tests:** Unit tests cover essential functions of the package
- [ x] **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines

- [x ] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

---

### Review Comments

Parzer package is a good and useful solution to parse the geographical coordinates in a lot of formats. We can save time using this tool.
Like contributor and user , I think it's necessary to improve the documentation and I suggest the implementation of more study cases.

Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient?
Maybe could be useful to add at least one example with South Latitude to see the negative numbers.

Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?
The package is coherent in the naming of the variables but it's a little ambivalent.
