### 2019-11-20 #341 @anadiedrichs
Thanks to @mvickm and @brunj7 for your valuable review. You gave great feedback which can help to improve the quality of the parzer package.
Let us wait & hear the @sckott's answer to chat on how to address the changes.
