### 2020-01-18 ropensci/parzer#17 @sckott
consider dropping Rcpp dependency
brought up in a review. not sure if possible or not, would need help on this
