### 2020-01-18 #341 @sckott
some clarification questions on some of the issues raised for the reviewers:

@brunj7 :

> Maybe `geographic coordinates` or `latitude and longitude coordinates`? I personally prefer the former because the use of geographic could also be referring to geographic vs projected coordinate systems (not covered by parzer by scope definition).

Do you mean "latter"? former would be `geographic coordinates`, correct?

> Is there a way to reduce Rccp dependencies?

Maybe? I'm a bit of a c++ noob, so if someone wants to help me do that i'm happy to do so. Is your main problem with it the extra dependency? performance? something else? Issue opened: https://github.com/ropenscilabs/parzer/issues/17

@mvickm :

> The package is coherent in the naming of the variables but it's a little ambivalent.

Can you explain in more detail what you mean by "ambivalent"?
