### 2020-01-18 674eb1a
update docs for new roxygen2 version; gitignore notes file
