+++ man/parse_lon.Rd
@@ -66,10 +66,10 @@ parse_lon("N 39 21.440") # => 39.35733
 parse_lon("S 56 1.389") # => -56.02315
 
 parse_lon("N40°25’5.994") # => 40.41833
-parse_lon("40° 25´ 5.994\\" N") # => 40.41833
+parse_lon("40° 25´ 5.994\" N") # => 40.41833
 parse_lon("40:25:6N")
 parse_lon("40:25:5.994N")
-parse_lon("40d 25’ 6\\" N")
+parse_lon("40d 25’ 6\" N")
 
 # user specfied format
 # \%C, \%c, \%H \%h \%D, \%d, \%M, \%m, \%S, and \%s
