+++ DESCRIPTION
@@ -13,7 +13,6 @@ BugReports: https://github.com/ropenscilabs/parzer/issues
 Imports:
     Rcpp (>= 1.0.2)
 Suggests:
-    roxygen2 (>= 6.1.1),
     testthat,
     randgeo,
     knitr,
@@ -23,4 +22,4 @@ SystemRequirements: C++11
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
 Encoding: UTF-8
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.2
