### 2020-01-18 1ef9db7
#19 added a use cases vignette skeleton
