+++ vignettes/use_cases.Rmd
@@ -1,0 +1,61 @@
+---
+title: "parzer use cases"
+author: "Scott Chamberlain"
+date: "`r Sys.Date()`"
+output:
+  html_document:
+    toc: true
+    toc_float: true
+    theme: readable
+vignette: >
+  %\VignetteIndexEntry{parzer use cases}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r echo=FALSE}
+knitr::opts_chunk$set(
+  comment = "#>",
+  collapse = TRUE,
+  warning = FALSE,
+  message = FALSE
+)
+```
+
+## Load parzer
+
+```{r}
+library("parzer")
+```
+
+## Use case: xxxx
+
+sasdfasdf
+
+```{r}
+"xxx"
+```
+
+## Use case: xxxx
+
+sasdfasdf
+
+```{r}
+"xxx"
+```
+
+## Use case: xxxx
+
+sasdfasdf
+
+```{r}
+"xxx"
+```
+
+## Use case: xxxx
+
+sasdfasdf
+
+```{r}
+"xxx"
+```
