### 2020-01-18 9dd8143
fix #15 add bit to readme about similar fxns in other pkgs
