### 2020-01-18 45908ce
fix #12 - always make lon first parameter if both lon and lat in a fxn definition bump version
