+++ tests/testthat/test-parse_hemisphere.R
@@ -2,13 +2,13 @@ context("parse_hemisphere")
 
 test_that("parse_hemisphere works", {
   # NE
-  ne <- parse_hemisphere("45N54.2356", "74.123E")
+  ne <- parse_hemisphere("74.123E", "45N54.2356")
   # NW
-  nw <- parse_hemisphere("40.4183318", "-120")
+  nw <- parse_hemisphere("-120", "40.4183318")
   # SW
-  sw <- parse_hemisphere("-40.4183318", "-120")
+  sw <- parse_hemisphere("-120", "-40.4183318")
   # SE
-  se <- parse_hemisphere("-40.4183318", "120")
+  se <- parse_hemisphere("120", "-40.4183318")
 
   for (i in c(ne, nw, sw, se)) expect_is(i, "character")
   for (i in c(ne, nw, sw, se)) expect_equal(nchar(i), 2)
@@ -16,17 +16,17 @@ test_that("parse_hemisphere works", {
   # bad values
   ## one
   expect_equal(
-    suppressWarnings(parse_hemisphere("-240.4183318", "120")), "E")
+    suppressWarnings(parse_hemisphere("120", "-240.4183318")), "E")
   expect_equal(
-    suppressWarnings(parse_hemisphere("-40.4183318", "420")), "S")
+    suppressWarnings(parse_hemisphere("420", "-40.4183318")), "S")
   ## both
   expect_equal(
-    suppressWarnings(parse_hemisphere("-240.4183318", "-200")), "")
+    suppressWarnings(parse_hemisphere("-200", "-240.4183318")), "")
 })
 
 test_that("parse_hemisphere - fails well", {
-  expect_error(parse_hemisphere(), "argument \"lat\" is missing")
-  expect_error(parse_hemisphere(""), "argument \"lon\" is missing")
-  expect_error(parse_hemisphere(mtcars), "lat must be of class")
-  expect_error(parse_hemisphere("", mtcars), "lon must be of class")
+  expect_error(parse_hemisphere(), "argument \"lon\" is missing")
+  expect_error(parse_hemisphere(""), "argument \"lat\" is missing")
+  expect_error(parse_hemisphere(mtcars), "lon must be of class")
+  expect_error(parse_hemisphere("", mtcars), "lat must be of class")
 })
