+++ R/zzz.R
@@ -7,9 +7,9 @@ assert <- function(x, y) {
   }
 }
 
-lint_inputs <- function(lat = NULL, lon = NULL, format) {
-  assert(lat, c("character", "numeric", "integer"))
+lint_inputs <- function(lon = NULL, lat = NULL, format) {
   assert(lon, c("character", "numeric", "integer"))
+  assert(lat, c("character", "numeric", "integer"))
   assert(format, "character")
 }
 
