### 2020-01-18 87eab3e
rbuildignore notes; contributing, no email option
