### 2020-01-18 aa83939
fix #16 change one letter math dms methods to have pz_ prefix
