### 2020-01-21 e419c3d
fix #14 add more examples using S latitudes throughout pkg
