+++ R/dms-fxns.R
@@ -24,17 +24,17 @@ dms_helper <- function(lon = NULL, lat = NULL) {
 #' to nonsense.
 #' @examples
 #' # extract parts of a coordinate value
-#' pz_degree(45.23323)
-#' pz_minute(45.23323)
-#' pz_second(45.23323)
+#' pz_degree(-45.23323)
+#' pz_minute(-45.23323)
+#' pz_second(-45.23323)
 #'
 #' pz_degree(lon = 178.23423)
 #' pz_minute(lon = 178.23423)
 #' pz_second(lon = 178.23423)
 #'
-#' pz_degree(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))
-#' pz_minute(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))
-#' pz_second(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))
+#' pz_degree(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
+#' pz_minute(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
+#' pz_second(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
 #'
 #' # invalid
 #' pz_degree(445.23323)
@@ -44,8 +44,8 @@ dms_helper <- function(lon = NULL, lat = NULL) {
 #' pz_m(44)
 #' pz_s(3)
 #' pz_d(31) + pz_m(44)
-#' pz_d(31) - pz_m(44)
-#' pz_d(31) + pz_m(44) + pz_s(59)
+#' pz_d(-31) - pz_m(44)
+#' pz_d(-31) + pz_m(44) + pz_s(59)
 #' pz_d(31) - pz_m(44) + pz_s(59)
 #' pz_d(-121) + pz_m(1) + pz_s(33)
 #' unclass(pz_d(31) + pz_m(44) + pz_s(59))
