+++ README.Rmd
@@ -46,7 +46,9 @@ parse_lat("45N54.2356")
 parse_lat("-45.98739874")
 parse_lat("40.123°")
 parse_lat("40.123N")
+parse_lat("40.123S")
 parse_lat("N45 04.25764")
+parse_lat("S45 04.25764")
 
 # bad values -> NaN
 parse_lat("191.89")
@@ -62,9 +64,11 @@ longitudes
 
 ```{r warning=TRUE}
 parse_lon("45W54.2356")
+parse_lon("45E54.2356")
 parse_lon("-45.98739874")
 parse_lon("40.123°")
 parse_lon("74.123W")
+parse_lon("74.123E")
 parse_lon("W45 04.25764")
 
 # bad values
