+++ README.md
@@ -55,8 +55,12 @@ parse_lat("40.123°")
 #> [1] 40.123
 parse_lat("40.123N")
 #> [1] 40.123
+parse_lat("40.123S")
+#> [1] -40.123
 parse_lat("N45 04.25764")
 #> [1] 45.07096
+parse_lat("S45 04.25764")
+#> [1] -45.07096
 
 # bad values -> NaN
 parse_lat("191.89")
@@ -80,12 +84,16 @@ longitudes
 ```r
 parse_lon("45W54.2356")
 #> [1] -45.90393
+parse_lon("45E54.2356")
+#> [1] 45.90393
 parse_lon("-45.98739874")
 #> [1] -45.9874
 parse_lon("40.123°")
 #> [1] 40.123
 parse_lon("74.123W")
 #> [1] -74.123
+parse_lon("74.123E")
+#> [1] 74.123
 parse_lon("W45 04.25764")
 #> [1] -45.07096
 
