+++ R/parse_lat_lon.R
@@ -1,4 +1,4 @@
-#' parse latitude and longitude
+#' parse longitude and latitude
 #'
 #' @export
 #' @param lon (character/numeric/integer) one or more longitude values
@@ -6,16 +6,16 @@
 #' @details length(lon) == length(lat)
 #' @return data.frame, with columns lon, lat
 #' @examples
-#' parse_lat_lon(-120.43, 49.12)
-#' parse_lat_lon(-120.43, 93)
-#' parse_lat_lon(-190, 49.12)
-#' parse_lat_lon(240, 49.12)
-#' parse_lat_lon(-190, 92)
+#' parse_lon_lat(-120.43, 49.12)
+#' parse_lon_lat(-120.43, 93)
+#' parse_lon_lat(-190, 49.12)
+#' parse_lon_lat(240, 49.12)
+#' parse_lon_lat(-190, 92)
 #' # many
 #' lons <- c("45W54.2356", "181", 45, 45.234234, "-45.98739874")
 #' lats <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
-#' parse_lat_lon(lons, lats)
-parse_lat_lon <- function(lon, lat) {
+#' parse_lon_lat(lons, lats)
+parse_lon_lat <- function(lon, lat) {
   lint_inputs(lon, lat, "")
   stopifnot(length(lon) == length(lat))
   data.frame(
