+++ tests/testthat/test-parse_lat_lon.R
@@ -1,7 +1,7 @@
-context("parse_lat_lon")
+context("parse_lon_lat")
 
-test_that("parse_lat_lon works", {
-  aa <- parse_lat_lon("45W54.2356", "45N54.2356")
+test_that("parse_lon_lat works", {
+  aa <- parse_lon_lat("45W54.2356", "45N54.2356")
 
   expect_is(aa, "data.frame")
   expect_equal(NCOL(aa), 2)
@@ -11,10 +11,10 @@ test_that("parse_lat_lon works", {
   expect_type(aa$lon, "double")
 })
 
-test_that("parse_lat_lon - fails well", {
-  expect_error(parse_lat_lon(), "argument \"lon\" is missing")
-  expect_error(parse_lat_lon(""), "argument \"lat\" is missing")
-  expect_error(parse_lat_lon(mtcars), "lon must be of class")
-  expect_error(parse_lat_lon("", mtcars), "lat must be of class")
-  # expect_error(parse_lat_lon("", "", 5), "format must be of class character")
+test_that("parse_lon_lat - fails well", {
+  expect_error(parse_lon_lat(), "argument \"lon\" is missing")
+  expect_error(parse_lon_lat(""), "argument \"lat\" is missing")
+  expect_error(parse_lon_lat(mtcars), "lon must be of class")
+  expect_error(parse_lon_lat("", mtcars), "lat must be of class")
+  # expect_error(parse_lon_lat("", "", 5), "format must be of class character")
 })
