+++ NAMESPACE
@@ -7,8 +7,8 @@ S3method("/",pz)
 S3method(print,pz)
 export(parse_hemisphere)
 export(parse_lat)
-export(parse_lat_lon)
 export(parse_lon)
+export(parse_lon_lat)
 export(parse_parts_lat)
 export(parse_parts_lon)
 export(pz_d)
