### 2020-01-21 bd696c0
ordering of lon/lat words in parse_hemisphere man file
