### 2020-01-21 6ff2c25
fix #18 add additional warning to fxns that handle both lon and lat: warn users to check if they inverted lon and lat when they input data
