+++ tests/testthat/test-parse_lat_lon.R
@@ -16,5 +16,8 @@ test_that("parse_lon_lat - fails well", {
   expect_error(parse_lon_lat(""), "argument \"lat\" is missing")
   expect_error(parse_lon_lat(mtcars), "lon must be of class")
   expect_error(parse_lon_lat("", mtcars), "lat must be of class")
-  # expect_error(parse_lon_lat("", "", 5), "format must be of class character")
+
+  # src/ error
+  expect_warning(parse_lon_lat(45, 190), "not within -90")
+  expect_warning(parse_lon_lat(45, 190), "check that you did not invert")
 })
