### 2020-01-22 b852014
#19 vignette added with a use case
