### 2020-01-22 a9ddd6e
#19 check if need to install pkgs in use cases vignette
