### 2020-02-03 #341 @brunj7
Hi @sckott

Thank you for addressing my comments!

To clarify my comment on the short description, I was suggesting to use: "parzer parses messy **geographic** coordinates". I think it helps to set the scope (vs coordinates alone)

Re: Rccp dependancies -- my comment was motivated by the fact that when I was playing around and reinstalling the package several times at some point the install failed. I could never reproduce the problem and restarting the session solved it. In the aftermath of this, I was wondering if reducing C dependancies could help to avoid any potential problem in the future (I could not reproduce the problem with this new version neither). This being said, my suggestion was more aiming at helping you to think about any ways to simplify dependancies, not at a request to do so. Unfortunately I am not knowledgeable in C and can not help.

@anadiedrichs I have reviewed the new version of `parzer` and reran all the tests on it. Everything looks good.  The package is good to go in my opinion.
