### 2020-02-04 #341 @sckott
thanks @brunj7

- Okay, i will fix that language to **geographic coordinates**
- Rcpp: okay, thanks for the explanation. there may be something with the Makevars file that I need to fix or so, i'll ask about that
