### 2020-02-10 #341 @mvickm
HI @sckott ! When I said "ambivalent", I was reffering about the naming of the variables. I mean, for the users  maybe could be a little confused using the same name. But I saw you have changed the name and add the prefix _pz. So for me is resolved, Thanks for consider the comments.
