### 2020-02-10 #341 @sckott
thanks @mvickm for the clarification

@anadiedrichs I need to address some wording changes from @brunj7 comments, and then I will have addressed all comments - will let you know when that's done
