### 2020-02-14 a02e475
makefile tweaks
