### 2020-02-14 81f935d
change wording in pkg to geographic coordinates from coordinates alone
