+++ README.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 [![Build Status](https://travis-ci.com/ropenscilabs/parzer.svg?branch=master)](https://travis-ci.com/ropenscilabs/parzer)
 [![codecov.io](https://codecov.io/github/ropenscilabs/parzer/coverage.svg?branch=master)](https://codecov.io/github/ropenscilabs/parzer?branch=master)
 
-`parzer` parses messy coordinates
+`parzer` parses messy geographic coordinates
 
 You may get data from a published study or a colleague, and the coordinates
 may be in some messy format that you'd like to clean up to e.g., have 
