### 2020-02-14 944ff7e
fill out description field in DESCRIPTION file add codemeta.json
