### 2020-02-14 5573539
add check windows to makefile
