+++ Makefile
@@ -22,6 +22,9 @@ check: build
 	@rm -f `ls -1tr ${PACKAGE}*gz | tail -n1`
 	@rm -rf ${PACKAGE}.Rcheck
 
+check_windows:
+	${RSCRIPT} -e "devtools::check_win_devel(); devtools::check_win_release()"
+
 eg:
 	${RSCRIPT} -e "devtools::run_examples()"
 
