### 2020-02-14 #341 @sckott
@anadiedrichs Okay, I've fixed the last thing raise about the wording "geographic coordinates". All done on my end
