### 2020-03-03 0af62bc
change all ropenscilabs to ropensci; add ropensci footer, ropensci review badge
