+++ .github/CONTRIBUTING.md
@@ -2,16 +2,16 @@
 
 ### Bugs?
 
-* Submit an issue on the [Issues page](https://github.com/ropenscilabs/parzer/issues)
+* Submit an issue on the [Issues page](https://github.com/ropensci/parzer/issues)
 
 ### Code contributions
 
 * Fork this repo to your Github account
 * Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/<yourgithubusername>/parzer.git`
-* Make sure to track progress upstream (i.e., on our version of `parzer` at `ropenscilabs/parzer`) by doing `git remote add upstream https://github.com/ropenscilabs/parzer.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
+* Make sure to track progress upstream (i.e., on our version of `parzer` at `ropensci/parzer`) by doing `git remote add upstream https://github.com/ropensci/parzer.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step
 * Make your changes (bonus points for making changes on a new feature branch)
 * Please do write a test(s) for your changes if they affect code and not just docs
 * Push up to your account
-* Submit a pull request to home base at `ropenscilabs/parzer`
+* Submit a pull request to home base at `ropensci/parzer`
 
 ### Also, check out our [discussion forum](https://discuss.ropensci.org)
