### 2020-03-03 ad8f424
reviewer to description file; add my ORCID; add ropensci as funder
