+++ DESCRIPTION
@@ -9,8 +9,13 @@ Description: Parse geographic coordinates from various formats
     and seconds. This package has been peer-reviewed by rOpenSci.
 Version: 0.0.9.99
 Authors@R: c(
-		person("Scott", "Chamberlain", role = c("aut", "cre"),
-            email = "sckott@protonmail.com")
+	person("Scott", "Chamberlain", role = c("aut", "cre"),
+        email = "sckott@protonmail.com",
+        comment = c(ORCID = "0000-0003-1444-9135")),
+    person("Julien", "Brun", role = "rev",
+        comment = "Julien Brun reviewed the package for rOpenSci,
+        see https://github.com/ropensci/onboarding/issues/341"),
+    person("rOpenSci", role = "fnd", comment = "https://ropensci.org")
 	)
 License: MIT + file LICENSE
 URL: https://github.com/ropensci/parzer (devel)
