### 2020-03-03 #341 @sckott
Thanks very much @anadiedrichs for your work - and thanks to the reviewers @brunj7 and @mvickm

I've done all of those things:
- transferred from ropenscilabs to ropensci
- added ropensci footer image
- fixed all links
- the docs page is still the ropenscilabs page, but i imagine that will update soon
- added the review mention in DESCRIPTION
- no badge links broken as far as I know
- already have a codemeta.json file

I added one of the reviewers to the DESCRIPTION - @mvickm i don't know your name, and not on your github profile

i'd love to do a tech post if okay with stef.
