### 2020-03-03 ff41301
fix some travis file config options [skip appveyor]
