+++ .travis.yml
@@ -1,9 +1,8 @@
 language: r
 cache: packages
-sudo: required
-dist: trusty
+os: linux
 
-matrix:
+jobs:
   include:
     - os: linux
       r: release
