### 2020-03-03 6d95176
single quote ropensci in description file
