### 2020-03-03 508037c
make one eg run faster to avoid cran eg time warning
