### 2020-03-03 a3a5acc
add news and cran comments file; bump to v0.1, update codemeta.json
