+++ cran-comments.md
@@ -1,0 +1,17 @@
+## Test environments
+
+* local OS X install, R 3.6.3 RC
+* ubuntu 16.04 (on travis-ci), R 3.6.2
+* win-builder (devel and release)
+
+## R CMD check results
+
+0 errors | 0 warnings | 0 notes
+
+---
+
+This is a new release. I have read and agree to the the
+CRAN policies at https://cran.r-project.org/web/packages/policies.html
+
+Thanks!
+Scott Chamberlain
