+++ NEWS.md
@@ -1,0 +1,6 @@
+parzer 0.1.0
+============
+
+### NEW FEATURES
+
+* Released to CRAN.
