+++ DESCRIPTION
@@ -6,9 +6,8 @@ Description: Parse geographic coordinates from various formats
     their parts (degree, minutes, seconds); calculate hemisphere
     from coordinates; pull out individually degrees,
     minutes, or seconds; add and subtract degrees, minutes,
-    and seconds. This package has been peer-reviewed by
-    'rOpenSci'.
-Version: 0.0.9.99
+    and seconds.
+Version: 0.1.0
 Authors@R: c(
 	person("Scott", "Chamberlain", role = c("aut", "cre"),
         email = "sckott@protonmail.com",
