### 2020-03-04 #341 @stefaniebutland
> i'd love to do a tech post if okay with stef.

🤔 ... yes please!
When do you think you'd have it ready - I can advise on publication date.
