### 2020-03-04 #341 @sckott
want to get on cran first. Getting it past cran could take a few days or who knows, maybe months, years even (months is possible).
