+++ DESCRIPTION
@@ -13,9 +13,13 @@ Authors@R: c(
         email = "sckott@protonmail.com",
         comment = c(ORCID = "0000-0003-1444-9135")),
     person("Julien", "Brun", role = "rev",
-        comment = "Julien Brun reviewed the package for rOpenSci,
+        comment = "Julien Brun reviewed the package,
         see https://github.com/ropensci/onboarding/issues/341"),
-    person("rOpenSci", role = "fnd", comment = "https://ropensci.org")
+    person("Maria", "Munafó", role = "rev",
+        comment = "Maria Munafó reviewed the package,
+        see https://github.com/ropensci/onboarding/issues/341"),
+    person("rOpenSci", role = "fnd",
+        comment = "https://ropensci.org")
 	)
 License: MIT + file LICENSE
 URL: https://github.com/ropensci/parzer (devel)
@@ -33,6 +37,7 @@ SystemRequirements: C++11
 VignetteBuilder: knitr
 Roxygen: list(markdown = TRUE)
 Encoding: UTF-8
+Language: en-US
 RoxygenNote: 7.0.2
 X-schema.org-applicationCategory: Geospatial
 X-schema.org-keywords: geospatial, data, latitude, longitude, parser, coordinates
