+++ man/parse_lat.Rd
@@ -67,8 +67,4 @@ parse_lat("40° 25´ 5.994\" N") # => 40.41833
 parse_lat("40:25:6N")
 parse_lat("40:25:5.994N")
 parse_lat("40d 25’ 6\" N")
-
-# user specfied format
-# \%C, \%c, \%H \%h \%D, \%d, \%M, \%m, \%S, and \%s
-# parse_lat("40.255994", "")
 }
