### 2020-03-13 8c75faf
fix #20 fix problem in parse_lon, wasnt catching number followed with [numbers]e[numbers]
