+++ tests/testthat/test-parse_lon.R
@@ -68,7 +68,8 @@ invalid_formats <- c(
   "60.1° N",
   "60.1° S",
   "N60.1",
-  "S60.1"
+  "S60.1",
+  "-45.23232e24"
   # "-40.4183318, 12.345, 74.6411133"
 )
 
