+++ tests/testthat/test-parse_lat.R
@@ -70,7 +70,8 @@ invalid_formats <- c(
   "60.1° W",
   "60.1° E",
   "W60.1",
-  "E60.1"
+  "E60.1",
+  "-45.23232e24"
   # "-40.4183318, 12.345, 74.6411133"
 )
 
