+++ R/parse_lon.R
@@ -37,23 +37,23 @@
 #' parse_lon(85:94)
 #'
 #' # different formats
-#' parse_lon("40.4183318 N")
-#' parse_lon("40.4183318 S")
+#' parse_lon("40.4183318 E")
+#' parse_lon("40.4183318 W")
 #' parse_lon("40 25 5.994") # => 40.41833
 #'
 #' parse_lon("40.4183318W")
 #' parse_lon("W40.4183318")
-#' parse_lon("E40.4183318S")
-#' parse_lon("N40.4183318")
+#' parse_lon("E40.4183318")
+#' parse_lon("40.4183318E")
 #'
-#' parse_lon("N 39 21.440") # => 39.35733
-#' parse_lon("S 56 1.389") # => -56.02315
+#' parse_lon("E 39 21.440") # => 39.35733
+#' parse_lon("W 56 1.389") # => -56.02315
 #'
-#' parse_lon("N40°25’5.994") # => 40.41833
-#' parse_lon("40° 25´ 5.994\" N") # => 40.41833
-#' parse_lon("40:25:6N")
-#' parse_lon("40:25:5.994N")
-#' parse_lon("40d 25’ 6\" N")
+#' parse_lon("E40°25’5.994") # => 40.41833
+#' parse_lon("40° 25´ 5.994\" E") # => 40.41833
+#' parse_lon("40:25:6E")
+#' parse_lon("40:25:5.994E")
+#' parse_lon("40d 25’ 6\" E")
 parse_lon <- function(lon, format = NULL) {
   assert(lon, c("numeric", "integer", "character"))
   assert(format, "character")
