### 2020-03-13 b7da440
tweaks to authors list, update cran comments
