+++ cran-comments.md
@@ -13,5 +13,8 @@
 This is a new release. I have read and agree to the the
 CRAN policies at https://cran.r-project.org/web/packages/policies.html
 
+This is a re-submission of the first version of this package, removing 
+commented out examples (all examples runs now), and fixing authors in DESCRIPTION file.
+
 Thanks!
 Scott Chamberlain
