+++ cran-comments.md
@@ -1,6 +1,6 @@
 ## Test environments
 
-* local OS X install, R 3.6.3 RC
+* local OS X install, R 3.6.3 Patched
 * ubuntu 16.04 (on travis-ci), R 3.6.2
 * win-builder (devel and release)
 
