+++ DESCRIPTION
@@ -28,7 +28,7 @@ Authors@R: c(
 	)
 License: MIT + file LICENSE
 URL: https://github.com/ropensci/parzer (devel)
-  https://docs.ropensci.org/parzer/ (docs)
+  https://docs.ropensci.org/parzer (docs)
 BugReports: https://github.com/ropensci/parzer/issues
 Imports:
     Rcpp (>= 1.0.2)
