### 2020-03-20 8a769de
#21 add symbol to scrub() hacky fxn for symbol from twitter conversation start test file for utility fxns;
