### 2020-03-20 bc4f3b8
fix #22 dont build vignettes on cran
