### 2020-03-20 e339ec1
skip nearly all tests on cran
