+++ tests/testthat/test-parse_lat_lon.R
@@ -1,6 +1,7 @@
 context("parse_lon_lat")
 
 test_that("parse_lon_lat works", {
+  skip_on_cran()
   aa <- parse_lon_lat("45W54.2356", "45N54.2356")
 
   expect_is(aa, "data.frame")
@@ -12,6 +13,7 @@ test_that("parse_lon_lat works", {
 })
 
 test_that("parse_lon_lat - fails well", {
+  skip_on_cran()
   expect_error(parse_lon_lat(), "argument \"lon\" is missing")
   expect_error(parse_lon_lat(""), "argument \"lat\" is missing")
   expect_error(parse_lon_lat(mtcars), "lon must be of class")
