+++ tests/testthat/test-parse_lon.R
@@ -1,6 +1,7 @@
 context("parse_lon")
 
 test_that("parse_lon works", {
+  skip_on_cran()
   aa <- parse_lon("45W54.2356")
 
   expect_is(aa, "numeric")
@@ -36,6 +37,7 @@ test_lons <- c(
 )
 
 test_that("parse_lon works: run through test_lons", {
+  skip_on_cran()
   out <- data.frame(input = test_lons, res = NA_real_,
     stringsAsFactors = FALSE)
   for (i in seq_along(test_lons)) {
@@ -46,6 +48,7 @@ test_that("parse_lon works: run through test_lons", {
 })
 
 test_that("parse_lon - fails well", {
+  skip_on_cran()
   expect_error(parse_lon(), "argument \"lon\" is missing")
   expect_error(parse_lon(mtcars), "lon must be of class")
   expect_error(parse_lon("", 5), "format must be of class character")
@@ -75,6 +78,7 @@ invalid_formats <- c(
 
 # res column should all give NaN
 test_that("parse_lon works: invalid formats fail as expected", {
+  skip_on_cran()
   out <- data.frame(input = invalid_formats, res = NA_real_,
                     stringsAsFactors = FALSE)
   for (i in seq_along(invalid_formats)) {
