### 2020-03-20 a16315d
dont run any egs on cran
