+++ R/parse_parts.R
@@ -11,7 +11,7 @@
 #'
 #' NA/NaN given upon error
 #'
-#' @examples
+#' @examples \dontrun{
 #' parse_parts_lon("45E54.2356")
 #' parse_parts_lon("140.4183318")
 #' parse_parts_lon("174.6411133")
@@ -32,6 +32,7 @@
 #' x <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
 #' parse_parts_lat(x)
 #' system.time(parse_parts_lat(rep(x, 10^2)))
+#' }
 
 #' @export
 #' @rdname parse_parts
