+++ R/parse_lat_lon.R
@@ -6,7 +6,7 @@
 #' @details length(lon) == length(lat)
 #' @return data.frame, with columns lon, lat. on an invalid values, an `NA`
 #' is returned. In addition, warnings are thrown on invalid values
-#' @examples
+#' @examples \dontrun{
 #' parse_lon_lat(-120.43, 49.12)
 #' parse_lon_lat(-120.43, 93)
 #' parse_lon_lat(-190, 49.12)
@@ -16,6 +16,7 @@
 #' lons <- c("45W54.2356", "181", 45, 45.234234, "-45.98739874")
 #' lats <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
 #' parse_lon_lat(lons, lats)
+#' }
 parse_lon_lat <- function(lon, lat) {
   lint_inputs(lon, lat, "")
   stopifnot(length(lon) == length(lat))
