+++ man/parse_lon.Rd
@@ -32,6 +32,7 @@ see \url{https://en.cppreference.com/w/cpp/error/invalid_argument}
 }
 
 \examples{
+\dontrun{
 parse_lon("")
 parse_lon("-181")
 parse_lon("-361")
@@ -70,4 +71,5 @@ parse_lon("40° 25´ 5.994\" E") # => 40.41833
 parse_lon("40:25:6E")
 parse_lon("40:25:5.994E")
 parse_lon("40d 25’ 6\" E")
+}
 }
