+++ man/parse_lon_lat.Rd
@@ -22,6 +22,7 @@ parse longitude and latitude
 length(lon) == length(lat)
 }
 \examples{
+\dontrun{
 parse_lon_lat(-120.43, 49.12)
 parse_lon_lat(-120.43, 93)
 parse_lon_lat(-190, 49.12)
@@ -31,4 +32,5 @@ parse_lon_lat(-190, 92)
 lons <- c("45W54.2356", "181", 45, 45.234234, "-45.98739874")
 lats <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
 parse_lon_lat(lons, lats)
+}
 }
