### 2020-03-20 0c0a9b4
update news and cran comments for next push, bump dev version
