+++ NEWS.md
@@ -1,3 +1,10 @@
+parzer 0.1.4
+============
+
+### MINOR IMPROVEMENTS
+
+* add support to internal function for additional degree like symbols (#21)
+
 parzer 0.1.0
 ============
 
