### 2020-03-26 d6c1208
run one eg in each man file
