+++ R/dms-fxns.R
@@ -22,7 +22,7 @@ dms_helper <- function(lon = NULL, lat = NULL) {
 #' but `/` and `*` are only exported with a stop message to say it's not
 #' supported; otherwise you'd be allow to divide degrees by minutes, leading
 #' to nonsense.
-#' @examples \dontrun{
+#' @examples
 #' # extract parts of a coordinate value
 #' pz_degree(-45.23323)
 #' pz_minute(-45.23323)
@@ -32,6 +32,7 @@ dms_helper <- function(lon = NULL, lat = NULL) {
 #' pz_minute(lon = 178.23423)
 #' pz_second(lon = 178.23423)
 #'
+#' \dontrun{
 #' pz_degree(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
 #' pz_minute(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
 #' pz_second(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 S"))
