+++ R/parse_hemisphere.R
@@ -12,9 +12,10 @@
 #' a zero length string.
 #' 
 #' Warnings are thrown on invalid values
-#' @examples \dontrun{
+#' @examples
 #' # NE
 #' parse_hemisphere("74.123E", "45N54.2356")
+#' \dontrun{
 #' # NW
 #' parse_hemisphere(-120, 40.4183318)
 #' # SW
