+++ man/parse_hemisphere.Rd
@@ -26,9 +26,9 @@ BEWARE: EXPERIMENTAL
 length(lon) == length(lat)
 }
 \examples{
-\dontrun{
 # NE
 parse_hemisphere("74.123E", "45N54.2356")
+\dontrun{
 # NW
 parse_hemisphere(-120, 40.4183318)
 # SW
