+++ man/parse_parts.Rd
@@ -27,8 +27,8 @@ NA/NaN given upon error
 parse coordinates into degrees, minutes and seconds
 }
 \examples{
-\dontrun{
 parse_parts_lon("45E54.2356")
+\dontrun{
 parse_parts_lon("140.4183318")
 parse_parts_lon("174.6411133")
 parse_parts_lon("-45.98739874")
