### 2020-03-26 0d4c7bb
attempt to fix clang-UBSAN errors with parse_parts_lon() fxn calls in vignette
