### 2020-03-26 63f25a9
update a man file, update vignette dae
