### 2020-03-26 aa8f813
bump to v0.1.4, update news and cran comments and codemeta.json
