### 2020-03-27 ac621e8
update cran comments and codemeta.son
