### 2020-03-31 1d73c9e
bump to dev version; add cran checks badge to readme
