### 2020-05-19 db533cd
translation of scrub in cpp + basic parsing of latlong strings
scrub_cpp does not perform well if system encoding is not UTF-8 and R does not convert the string to UTF-8 before passing to scrub_cpp.
R converts automatically in presence of some special characters (like ″) but not all (like º). If R does not convert, no regex match or replacement.
Conversion should be on the cpp side with ICU library functions.

No testing of the latlong string parser yet.
