### 2020-06-26 6bc89fc
Added control of the encoding of the inputs in scrub()
