### 2020-06-26 bb30c40
Update of parse_llstr that does the split and parse latitude and longitude
Takes a vector of strings with latitude and longitude and returns a data.frame
Update of the tests too.
